more docstrings...
It's not always easy to decide when a given term should go between ` `
in order to show up as code-style. Wherever it seemed obvious and
convenient I've added them. Avoided to use it where I wasn't sure, and
always kept in mind some balance between readability in source-code and
on Wiki.