rst directive parse shorthand and more compat
with the ```eval_rst style, blocks, certain directives such as 'sidebar'
failed with an error saying: "ERROR: The "sidebar" directive may not be
used within topics or body elements". The sidebar directive works using
this added style. The new style is: ```sidebar:: Sidebar title. The
double colon denotes that it is an rst directive and renders it
correctly. No code-block languages have double colons, so this is a safe
method to distinguish between code literals and rst directives.