Contrib Proposal: ASCII Map Builder
This is a command capable of taking a reference to a basic 2D ASCII map stored in the Evennia directory and  generating the rooms and exits  necessary to populate that world. The characters of the map are iterated over and compared to a list of trigger characters. When a match is found it triggers the corresponding instructions.

You then call the command in-game using the path to the module and the name of the variable holding the map string. The path you provide is relative to the evennia or your mygame folder.
    @mapbuilder <path.to.module.VARNAME>