Change priority to highest.
Currently the priority for the Autostructify is set to 1000.  I think it
was the intention to set it so that it runs before any other transforms.
However the ordering in docutils is confusing and setting it to 1
actually means it will be run first.

This was causing a problem for i18n in sphinx as it would not translate
anything in ```eval_rst`` blocks or ``` directive:: ``` blocks as these were
not converted to the node tree before gettext was meant to translate
them.