Added combat help, fixed oversight in CmdFight
Added a new command, 'CmdCombatHelp', which inherits from the default help command. Its behavior is identical to the normal 'help' command, except that it'll return a short list of combat commands when used in combat with no arguments. If there's a better way to implement this functionality, let me know, since it did involve copy-pasting a huge chunk of code from the Help command just to change one thing!

I also fixed a glaring oversight where using the 'fight' command while already in a fight would start a second fight.