Add a @spawn unit test
Tests the @spawn command and its variations.  The test is designed to
aslo check the properties of the spawned objects for some common/obvious
settings.

Test also adds a new prototypes file to be used with the Evennia unit
tester for any tests that require named prototypes.  The spawner test
requires testing named prototypes, so it is added in this change.