Fix an issue with adding an empty column to an EvTable
There seems to be an issue with adding an empty column (i.e. a column
with empty data) to an EvTable that has already been set up.  It seems
that the column with empty data is added with one extra row than the
rest of the table, and when a new call to add_rows() is made to EvTable,
the data in the column that was added previously is offset by 1.

This change fixes that by calculating the size of the new column to take
into account the headers BEFORE making the calculation on whether to
expand the column to match the table's size or not.