Batchcode Build.py for Tutorial World.
This is a draft of a proposed Batchcode Build.py for the Tutorial World.

I still need to make a second pass at the documentation, fix the bug where the Ghostly Apparitions think they own the place (not respecting the locks) and conduct more rigorous testing. Uploaded at this stage to get feedback on layout and other stylistic decisions.