Fixing bug in gametime module
Previously, gametime used the length of the current year and
current month to calculate the future date in the real_seconds_until
function. This caused a small bug when passing over a leap day from
a year without a leap day.

This change uses the datetime module to just increment the year
or month, sidestepping the issue.