Implemented suggested changes.
Removed scroll as a format option and replaced it with no_bars and invert.
Added header
Updated docstring to hidekeys and updated both docstring and code to change movedesc and hidedesc to movedescs and hidedescs
Removed ANSIljust since it only existed to support scroll

Not implemented while being discussed:
Still using the key of 'format' until either a new term is settled on or it is decided that the current term is all right.
Have not added header-, text-, or footer- format options. While I like the idea I think more discussion needs to be given to exactly what these options do, it sort of feels like we can always add another feature so I'd like to 'feature lock' at the current position and we can add them later, and most likely the effect of these can be easily implemented inside the menu itself (adding a formatting function to the menu and then simply saying 'text': special_format(text).