Add null checking when getting puppet from session
When account.execute_cmd() is called without an explicit session
argument for a server using MULTISESSION_MODE 2 or 3, it does not
know which session to use and may leave it None. Downstream, code
attempting to get the puppet from the session can fail with an
error because of the None session.

This change adds in checking for the None session, and sets the puppet
to None if there is no session defined.