HelpDetailView object cache
Created a cache for the object (target of the help request) in a HelpDetailView.
This reduces the call of DetailView.get_object by three or four times. Which also reduces the number of DetailView.get_queryset calls by the same amount.
Unit test created. To verify requesting a new object does not return the result from a previous request.