chore: convert try/except to explicit checks
Rather than causing an exception and handling it, this change explicitly checks the type of certain arguments and only converts strings to integers if needed.

Both this change and the previous implementation don't handle cases where the string passed is not numeric. However, I'm just getting familiar with the codebase, so this is possibly handled elsewhere.