fix: remove erroneous links from the documentation
Sphinx assumes that WORD.py should be a link to https://WORD.py .

This is not a problem in most of the documentation, since Sphinx won't
turn those python filenames into links when they are wrapped in
backticks, like `WORD.py`. Unfortunately, not always that was being
done, and so there were several wrong, broken links in the
documentation.

This patch wraps all the occurrences I've found of this case with
backticks, not only making the documentation more readible and
homogeneous, but more importantly getting rid of those unwanted links
in the generated HTML version of the documentation.