Fix bug adding duplicate components.
Adding a component to a ComponentHandler replaces the entry in
_loaded_components, but adds to some lists (signals and the
"component_names" attribute). The lists keep growing and cause
undesireable behaviors (duplicate signal handling, inability to remove
the components).

This change explicitly removes components with the same slot name
before insertion. It also results in new calls to at_removed() which
did not happen in previous component replacements but should have.