The tag “closed” or its translated version is now automatically added to the tag tuple if it has not already been specified in “settings.py”.
Current behavior: If the tag 'closed' is not set in
INGAME_REPORT_STATUS_TAGS, reports can no longer be closed, which seems
like a bug to me.

An alternative solution would be another setting in ‘settings.py’, such
as ‘INGAME_REPORT_STATUS_CLOSED_TAG’. This would also eliminate the need
for internationalization, as the tag could be set in the desired
language.